Add initial remote bootstrap script
The bootstrap script is a bash script that is put onto the Raspberry Pi by a
simple Ruby deploy script (needed in order to pass a password to ssh).

I'm using bish to write the bash script because I hate Bash syntax but don't
want to require any other scripting language to be preinstalled on the
Raspberry Pi.